refactor: remove ProcessesList component
