+++ tasklist/client/src/v1/ProcessesTab/index.tsx
@@ -6,7 +6,17 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {type InlineLoadingProps} from '@carbon/react';
+import {
+  Column,
+  Dropdown,
+  Grid,
+  Layer,
+  Link,
+  Search,
+  SkeletonPlaceholder,
+  Stack,
+  type InlineLoadingProps,
+} from '@carbon/react';
 import debounce from 'lodash/debounce';
 import {
   useLocation,
@@ -36,8 +46,43 @@ import {useForm} from 'v1/api/useForm.query';
 import {getProcessDisplayName} from 'v1/utils/getProcessDisplayName';
 import {useStartProcess} from 'v1/api/useStartProcess.mutation';
 import type {Process} from 'v1/api/types';
-import {ProcessesList} from 'common/processes/ProcessesList';
-import {START_FORM_FILTER_OPTIONS} from 'common/processes/ProcessesList/constants';
+import {
+  START_FORM_FILTER_OPTIONS,
+  type FilterOption,
+} from 'common/processes/constants';
+import {MultitenancyDropdown} from 'common/multitenancy/MultitenancyDropdown';
+import {C3EmptyState} from '@camunda/camunda-composite-components';
+import EmptyMessageImage from 'common/processes/empty-message-image.svg';
+import {ProcessTile} from 'common/processes/ProcessTile';
+import {getMultiModeProcessDisplayName} from 'common/processes/getMultiModeProcessDisplayName';
+import {useIsMultitenancyEnabled} from 'common/multitenancy/useIsMultitenancyEnabled';
+
+const FilterDropdown: React.FC<{
+  items: FilterOption[];
+  selected?: FilterOption;
+  onChange?: (option: FilterOption) => void;
+}> = ({items, selected, onChange}) => {
+  const {t} = useTranslation();
+
+  return (
+    <Dropdown
+      id="process-filters"
+      data-testid="process-filters"
+      className={styles.dropdown}
+      hideLabel
+      selectedItem={selected}
+      titleText={t('processesFilterDropdownLabel')}
+      label={t('processesFilterDropdownLabel')}
+      items={items}
+      itemToString={(item) => (item ? t(item.textKey) : '')}
+      onChange={(data) => {
+        if (data.selectedItem && onChange) {
+          onChange(data.selectedItem);
+        }
+      }}
+    />
+  );
+};
 
 type InlineLoadingStatus = NonNullable<InlineLoadingProps['status']>;
 
@@ -143,6 +188,39 @@ const ProcessesTab: React.FC = observer(() => {
       });
     },
   });
+  const processSearchProps: React.ComponentProps<typeof Search> = {
+    size: 'md',
+    placeholder: t('processesFilterFieldLabel'),
+    labelText: t('processesFilterFieldLabel'),
+    closeButtonLabelText: t('processesClearFilterFieldButtonLabel'),
+    value: searchValue,
+    onChange: (event) => {
+      setSearchValue(event.target.value);
+      debouncedNavigate(searchParams, {
+        name: 'search',
+        value: event.target.value,
+      });
+    },
+    disabled: isLoading,
+  } as const;
+  const startFormFilterDropdownProps: React.ComponentProps<
+    typeof FilterDropdown
+  > = {
+    items: START_FORM_FILTER_OPTIONS,
+    selected: startFormFilter,
+    onChange: (value) =>
+      debouncedNavigate(searchParams, {
+        name: 'hasStartForm',
+        value: value.searchParamValue ?? '',
+      }),
+  } as const;
+  const initialTenant = currentUser?.tenants.find(({tenantId}) =>
+    [searchParams.get('tenantId') ?? undefined, getStateLocally('tenantId')]
+      .filter((tenantId) => tenantId !== undefined)
+      .includes(tenantId),
+  );
+  const processes = data?.processes ?? [];
+  const {isMultitenancyEnabled} = useIsMultitenancyEnabled();
 
   useEffect(() => {
     if (match !== null) {
@@ -223,103 +301,221 @@ const ProcessesTab: React.FC = observer(() => {
     <main className={cn('cds--content', styles.splitPane)}>
       <NewProcessInstanceTasksPolling />
 
-      <ProcessesList
-        processes={data?.processes ?? []}
-        searchValue={searchValue}
-        isLoading={isLoading}
-        isFiltered={isFiltered}
-        onSearch={(event) => {
-          setSearchValue(event.target.value);
-          debouncedNavigate(searchParams, {
-            name: 'search',
-            value: event.target.value,
-          });
-        }}
-        initialTenant={currentUser?.tenants.find(({tenantId}) =>
-          [
-            searchParams.get('tenantId') ?? undefined,
-            getStateLocally('tenantId'),
-          ]
-            .filter((tenantId) => tenantId !== undefined)
-            .includes(tenantId),
-        )}
-        onTenantChange={(tenant) => {
-          updateSearchParams(searchParams, {
-            name: 'tenantId',
-            value: tenant,
-          });
-          storeStateLocally('tenantId', tenant);
-        }}
-        startFormFilterValue={startFormFilter}
-        onStartFormFilterChange={(value) =>
-          debouncedNavigate(searchParams, {
-            name: 'hasStartForm',
-            value: value.searchParamValue ?? '',
-          })
-        }
-        isStartButtonDisabled={instance !== null}
-        getUniqueId={(process) => process.id}
-        onStartProcess={(process: Process) => async () => {
-          setSelectedProcess(process);
-          const {bpmnProcessId} = process;
-          if (process.startEventFormId === null) {
-            setStartProcessStatus('active');
-            tracking.track({
-              eventName: 'process-start-clicked',
-            });
-            try {
-              await startProcess({
-                bpmnProcessId,
-                tenantId: selectedTenantId,
-              });
-            } catch (error) {
-              logger.error(error);
-              setStartProcessStatus('error');
-            }
-          } else {
-            navigate({
-              ...location,
-              pathname: pages.internalStartProcessFromForm(bpmnProcessId),
-            });
-          }
-        }}
-        onStartProcessError={(process: Process) => () => {
-          setSelectedProcess(null);
-          const displayName = getProcessDisplayName(process);
-          tracking.track({
-            eventName: 'process-start-failed',
-          });
-          setStartProcessStatus('inactive');
-          if (
-            getClientConfig().isMultiTenancyEnabled &&
-            selectedTenantId === undefined
-          ) {
-            notificationsStore.displayNotification({
-              isDismissable: false,
-              kind: 'error',
-              title: t('processesStartProcessFailedMissingTenant'),
-              subtitle: displayName,
-            });
-          } else {
-            notificationsStore.displayNotification({
-              isDismissable: false,
-              kind: 'error',
-              title: t('processesStartProcessFailed'),
-              subtitle: displayName,
-            });
-          }
-        }}
-        onStartProcessSuccess={() => {
-          setSelectedProcess(null);
-          setStartProcessStatus('inactive');
-        }}
-        selectedTenantId={selectedTenantId}
-        getStartProcessStatus={(process: Process) =>
-          selectedProcess?.bpmnProcessId === process.bpmnProcessId
-            ? startProcessStatus
-            : 'inactive'
-        }
-      />
+      <div className={styles.container}>
+        <Stack className={styles.content} gap={2}>
+          <div className={styles.searchContainer}>
+            <Stack className={styles.searchContainerInner} gap={6}>
+              <Grid narrow>
+                <Column sm={4} md={8} lg={16}>
+                  <Stack gap={4}>
+                    <h1>{t('headerNavItemProcesses')}</h1>
+                    <p>{t('processesSubtitle')}</p>
+                  </Stack>
+                </Column>
+              </Grid>
+              {isMultitenancyEnabled ? (
+                <Grid narrow>
+                  <Column
+                    className={styles.searchFieldWrapper}
+                    sm={4}
+                    md={8}
+                    lg={10}
+                  >
+                    <Search {...processSearchProps} />
+                  </Column>
+                  <Column
+                    className={styles.searchFieldWrapper}
+                    sm={2}
+                    md={4}
+                    lg={3}
+                  >
+                    <FilterDropdown {...startFormFilterDropdownProps} />
+                  </Column>
+                  <Column
+                    className={styles.searchFieldWrapper}
+                    sm={2}
+                    md={4}
+                    lg={2}
+                  >
+                    <MultitenancyDropdown
+                      initialSelectedItem={initialTenant}
+                      onChange={(tenant) => {
+                        updateSearchParams(searchParams, {
+                          name: 'tenantId',
+                          value: tenant,
+                        });
+                        storeStateLocally('tenantId', tenant);
+                      }}
+                    />
+                  </Column>
+                </Grid>
+              ) : (
+                <Grid narrow>
+                  <Column
+                    className={styles.searchFieldWrapper}
+                    sm={4}
+                    md={5}
+                    lg={10}
+                  >
+                    <Search {...processSearchProps} />
+                  </Column>
+                  <Column
+                    className={styles.searchFieldWrapper}
+                    sm={4}
+                    md={3}
+                    lg={5}
+                  >
+                    <FilterDropdown {...startFormFilterDropdownProps} />
+                  </Column>
+                </Grid>
+              )}
+            </Stack>
+          </div>
+
+          <div className={styles.processTilesContainer}>
+            <div className={styles.processTilesContainerInner}>
+              {!isLoading && processes.length === 0 ? (
+                <Layer>
+                  <C3EmptyState
+                    icon={
+                      isFiltered
+                        ? undefined
+                        : {path: EmptyMessageImage, altText: ''}
+                    }
+                    heading={
+                      isFiltered
+                        ? t('processesProcessNotFoundError')
+                        : t('processesProcessNotPublishedError')
+                    }
+                    description={
+                      <span data-testid="empty-message">
+                        {t('processesErrorBody')}
+                        <Link
+                          href="https://docs.camunda.io/docs/components/modeler/web-modeler/run-or-publish-your-process/#publishing-a-process"
+                          target="_blank"
+                          rel="noopener noreferrer"
+                          inline
+                          onClick={() => {
+                            tracking.track({
+                              eventName: 'processes-empty-message-link-clicked',
+                            });
+                          }}
+                        >
+                          {t('processesErrorBodyLinkLabel')}
+                        </Link>
+                      </span>
+                    }
+                  />
+                </Layer>
+              ) : (
+                <Grid narrow as={Layer}>
+                  {isLoading
+                    ? Array.from({length: 5}).map((_, index) => (
+                        <Column
+                          className={styles.processTileWrapper}
+                          sm={4}
+                          md={4}
+                          lg={5}
+                          key={index}
+                        >
+                          <SkeletonPlaceholder
+                            className={styles.tileSkeleton}
+                            data-testid="process-skeleton"
+                          />
+                        </Column>
+                      ))
+                    : processes.map((process, idx) => (
+                        <Column
+                          className={styles.processTileWrapper}
+                          sm={4}
+                          md={4}
+                          lg={5}
+                          key={process.id}
+                        >
+                          <ProcessTile
+                            process={process}
+                            displayName={getMultiModeProcessDisplayName(
+                              process,
+                            )}
+                            isFirst={idx === 0}
+                            isStartButtonDisabled={instance !== null}
+                            data-testid="process-tile"
+                            tenantId={selectedTenantId}
+                            onStartProcess={async () => {
+                              setSelectedProcess(process);
+                              const {bpmnProcessId} = process;
+                              if (process.startEventFormId === null) {
+                                setStartProcessStatus('active');
+                                tracking.track({
+                                  eventName: 'process-start-clicked',
+                                });
+                                try {
+                                  await startProcess({
+                                    bpmnProcessId,
+                                    tenantId: selectedTenantId,
+                                  });
+                                } catch (error) {
+                                  logger.error(error);
+                                  setStartProcessStatus('error');
+                                }
+                              } else {
+                                navigate({
+                                  ...location,
+                                  pathname:
+                                    pages.internalStartProcessFromForm(
+                                      bpmnProcessId,
+                                    ),
+                                });
+                              }
+                            }}
+                            onStartProcessError={() => {
+                              setSelectedProcess(null);
+                              const displayName =
+                                getProcessDisplayName(process);
+                              tracking.track({
+                                eventName: 'process-start-failed',
+                              });
+                              setStartProcessStatus('inactive');
+                              if (
+                                getClientConfig().isMultiTenancyEnabled &&
+                                selectedTenantId === undefined
+                              ) {
+                                notificationsStore.displayNotification({
+                                  isDismissable: false,
+                                  kind: 'error',
+                                  title: t(
+                                    'processesStartProcessFailedMissingTenant',
+                                  ),
+                                  subtitle: displayName,
+                                });
+                              } else {
+                                notificationsStore.displayNotification({
+                                  isDismissable: false,
+                                  kind: 'error',
+                                  title: t('processesStartProcessFailed'),
+                                  subtitle: displayName,
+                                });
+                              }
+                            }}
+                            onStartProcessSuccess={() => {
+                              setSelectedProcess(null);
+                              setStartProcessStatus('inactive');
+                            }}
+                            status={
+                              selectedProcess?.bpmnProcessId ===
+                              process.bpmnProcessId
+                                ? startProcessStatus
+                                : 'inactive'
+                            }
+                          />
+                        </Column>
+                      ))}
+                </Grid>
+              )}
+            </div>
+          </div>
+        </Stack>
+      </div>
 
       <FormModal
         processDisplayName={
