+++ tasklist/client/src/v1/ProcessesTab/styles.module.scss
@@ -17,3 +17,48 @@
   flex-direction: row;
   background: var(--cds-layer);
 }
+
+.container {
+  width: 100%;
+  overflow: auto;
+}
+
+.content {
+  height: fit-content;
+  flex: 1;
+}
+
+.processTilesContainer {
+  background: var(--cds-layer);
+}
+
+.searchContainer {
+  background: var(--cds-background);
+}
+
+.searchContainerInner {
+  max-width: 99em;
+  padding: var(--cds-spacing-06) var(--cds-spacing-08);
+}
+
+.processTilesContainerInner {
+  max-width: 99em;
+  padding: var(--cds-spacing-06) var(--cds-spacing-08);
+}
+
+.searchFieldWrapper {
+  padding: var(--cds-spacing-03) 0;
+}
+
+.dropdown {
+  width: 100%;
+}
+
+.processTileWrapper {
+  padding: var(--cds-spacing-03) 0;
+}
+
+.tileSkeleton {
+  width: 100%;
+  height: layout.to-rem(132px);
+}
